Added mechanisms to test functionality with deduplication module and without

+++ run_tests.sh
@@ -3,10 +3,19 @@
 export BOAR_CACHEDIR=`mktemp --tmpdir=/tmp/ -d "boar_tests_cache_XXXXX"`
 export BOAR_SERVER_CLI="`pwd`/boar"
 
+test -e run_tests.sh || { echo "This command must be executed in the boar installation top dir"; exit 1; }
+test -e rollingcs.so || { echo "ERROR: dedup module unavailable"; exit 1; }
+
+#
+# Test with deduplication
+#
+
 for unittest in tests/test_*.py blobrepo/tests/test_*.py; do
     echo "Excuting $unittest (cachedir $BOAR_CACHEDIR)"
     python $unittest || { echo "Unittest $unittest failed"; exit 1; }
 done
+
+
 rm -r $BOAR_CACHEDIR
 
 echo "Executing local macro tests"
